Housecleaning!
These packages were originally meant to hold a bunch of ridiculously overbroad utilities.  They don't belong here.  This repo is being trimmed down to the exec lib, and nothing else.

Also remove submodule to kr's pty library.  We had plans to use it a long time ago, but they haven't appeared yet.  We'll deal with this again when it's actually bound to some use.

Signed-off-by: Eric Myhre <hash@exultant.us>