Start work on the next generation of gosh.  This is going to be a large shake-up: most core components are going to become interfaces instead of concrete structures.
Here are the first of the major changes:

- process state is now a typed constant, and includes its own methods instead of heaping them onto the whole command observer struct
- the command observer system is now called `Proc` instead of `RunningCommand`
- the command observer system is now an interface instead stuck to a single concrete implementation
  - `ProcExec` is now just one implementation of `Proc`; you can make more!

The entire command templating system is, for the moment, dropped.  When it's revived, the "baking" behaviors will have become the exception, rather than the rule -- it's very rare (relatively) that you want to invoke the same command again and again with no args, which means requiring the nigh-invisible trailing `()` (or `Run()`) is lot more pain than it's worth.  Instead, we're going to look into ways to make the interface "DTRT" more when invoked with a series of string args that clearly meant to kick something off.

Tests are MISSING.  We're gonna get a clean start with those, too, using goconvey this time.

Note that I'd like to have placed the ProcExec implementation off in its own package, but if the main gosh package is to be able to invoke it without fuss, that'd result in import cycle issues, so it appears it's not meant to be.

Signed-off-by: Eric Myhre <hash@exultant.us>