New tests for ProcExec.
Most of the old coverage is restored... and it's much terser and more wonderful with goconvey.  Some groups of checks still todo.

Added some time bounding assertions as well -- obviously these are inherently racy, but the lower bound is safe, and the margin on the upper bound is so big I dearly hope it's reasonable.  And of course, the happy path is not slow, because that's how you murder your iteration time.

Signed-off-by: Eric Myhre <hash@exultant.us>