Speed up test.  Tricky.
When the sig handler exits. I'd *like* for this to cause the whole bash process to exit immediately.  Previously, it didn't.  That is now fixed.

This contained a multi-stage cluster... er, clusterchallenge:

1. By default, when bash receives a signal, it waits for the current task to complete, *then* disbatches the handler.
   Putting the sleep in a backgrounded subshell and then waiting for it fixes waiting on this, because the 'wait' builtin has magically different behavior in this situation.
2. Attempts to 'disown' before the wait would mean the wait wouldn't, well, wait; so that doesn't work.  Doing a disown before the exit in the handler *does* disown the job, but that turns out to be a red herring because...
3. If the process *running* this whole show is waiting for the stdout fd to be closed... that sleep call is holding it open, and that gumms up the whole works.

So testing the signal behavior and having output from it at the same time so we can trace the flow turns out to be nigh impossible to do correctly.  :/  Fortunately, that's more of a nice-to-have than critical to test correctness, so we're just gonna drop output collection entirely.

There are still other race conditions here.  That 100ms wait to be sure bash has finished setting up its trap is fairly disheartening (and it's one of the supershitty kinds of sleeps where the normal case is the full length nap).  We could fix it by having an output line that declares readiness, but, well... as we just discovered, that's wading into treacherous waters.  So, maybe some day if someone can work up the patience to debug all that, great; for now that sleep is so egregiously long that it works well enough in practice, and fortunately there aren't enough of these to dramaticaly slow iteration times.

Signed-off-by: Eric Myhre <hash@exultant.us>