Add a new Launcher which allows total customization of the exec.Cmd just before launch.
This is an escape valve for doing anything you need with the exec.Cmd -- while still being scoped to the use of the Exec Proc, so that it's quite clear from the API itself that other Proc implementations are not bound to allow such things.

You can use this to do linux specific things and other powerful nonsense that the shellish layer of gosh typically prevents you from doing.  Like any good escape valve, this API aims to be usable enough to make anything possible, but frictional enough to make you look for a better way.

Signed-off-by: Eric Myhre <hash@exultant.us>